+++ docs/changelog.rst
@@ -5,7 +5,9 @@ CHANGELOG
 2.11.2.dev0
 -------------------
 
-*
+**Bug fixes**
+
+* Improve performances of DEM computation for huge treks
 
 
 2.11.1 (2016-08-17)