+++ docs/changelog.rst
@@ -2,16 +2,19 @@
 CHANGELOG
 =========
 
-2.11.2.dev0
+2.11.2 (2016-09-15)
 -------------------
 
 **Bug fixes**
 
-* Improve performances of DEM computation for huge treks
 * Do not synchronize not-published treks with published but deleted parents
 * Allow to specify portal in touristic content parsers
 * Fix import of type1 in HebergementsSitraParser
 * Fix source and portal missing in shapefile exports
+
+**Performances**
+
+* Improve performances of DEM computation for huge treks
 
 
 2.11.1 (2016-08-17)